pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "scaffold-repo"
version = "0.1.0"
description = "Scaffold C/CMake repos from Jinja2 templates and enforce OSS hygiene."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "Andy Curtis", email = "contactandyc@gmail.com" }]
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
scaffold-repo = "scaffold_repo.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pre-commit>=3.0",
    "ruff>=0.4",
    "black>=24.0",
    "pathspec>=0.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scaffold_repo"]